Rename two files
